backend/requirements.txt
Django==4.2.7
djoser==2.3.0
djangorestframework==3.14.0
django-cors-headers==4.3.0
django-filter==23.3
django-allauth==0.57.0
django-rest-auth==0.9.5
channels==4.0.0
channels-redis==4.1.0
daphne==4.0.0
psycopg2-binary==2.9.9
Pillow==10.1.0
python-dotenv==1.0.0
djangorestframework-simplejwt==5.3.0
django-storages==1.14.2
boto3==1.28.62
gunicorn==21.2.0
whitenoise==6.5.0
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-redis==5.4.0
# Removed django-channels-webrtc==0.1.0 due to Python version compatibility issues
# Removed django-channels==1.0.0 due to Python version compatibility issues
# WebRTC will be implemented directly using:
aiortc==1.5.0
# For WebRTC signaling, we'll use the existing channels infrastructure
stripe==7.0.0
celery==5.3.4
redis==5.0.1
pytest==7.4.3
pytest-django==4.5.2
coverage==7.3.2
factory-boy==3.3.0
Faker==19.12.0
black==23.10.1
isort==5.12.0
flake8==6.1.0
mypy==1.6.1
